+++ satella/instrumentation/metrics/metric_types/__init__.py
@@ -1,9 +1,9 @@
 from .base import Metric, LeafMetric, RUNTIME, DISABLED, INHERIT, DEBUG, EmbeddedSubmetrics
 from .cps import ClicksPerTimeUnitMetric
 from .simple import IntegerMetric, FloatMetric
-from .percentile import PercentileMetric
+from .quantile import QuantileMetric
 from .registry import register_metric, METRIC_NAMES_TO_CLASSES
 
 __all__ = ['Metric', 'LeafMetric', 'EmbeddedSubmetrics', 'RUNTIME', 'DEBUG', 'INHERIT',
            'DISABLED', 'ClicksPerTimeUnitMetric', 'IntegerMetric', 'FloatMetric',
-           'PercentileMetric', 'register_metric', 'METRIC_NAMES_TO_CLASSES']
+           'QuantileMetric', 'register_metric', 'METRIC_NAMES_TO_CLASSES']